# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cocohub API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Cocohub Engineering",
      "email": "engineering@cocohub.app",
      "url": "https://cocohub.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://staging.cocohub.app/api",
      "description": "Staging environment"
    },
    {
      "url": "https://api.cocohub.app/api",
      "description": "Production environment"
    }
  ],
...
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "description": "Exchanges a valid refresh token for a new access token. The old refresh token is invalidated. Called automatically by the API client interceptor on 401 responses.",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Refresh token expired or invalid — user must log in again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "REFRESH_TOKEN_EXPIRED",
                    "message": "Session expired — please log in again"
                  },
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/auth/oauth/{provider}": {
      "post": {
        "tags": ["Authentication"],
        "summary": "OAuth login/register via social provider",
        "description": "Authenticates or registers a user via a social OAuth provider (Google, Apple, Facebook). The client obtains an authorization code from the provider and passes it here.",
        "operationId": "oauthLogin",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["google", "apple", "facebook"]
            },
            "example": "google"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `POST /auth/login` or `POST /auth/register`.\n\n**Example:**\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nTokens expire after 7 days. Use `POST /auth/refresh` to obtain a new token."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for server-to-server integrations. Contact engineering@cocohub.app to obtain a key."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": ["success", "error", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input data"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "example": {
                  "field": "email",
                  "issue": "Invalid format"
                }
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
...
      "UUID": {
        "type": "string",
        "format": "uuid",
        "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "The refresh token issued at login"
          }
        }
      },
...
      "OAuthCallbackRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string",
            "example": "oauth-authorization-code",
            "description": "OAuth 2.0 authorization code from the provider"
          },
          "redirectUri": {
            "type": "string",
            "format": "uri",
            "example": "cocohub://auth/callback"
          }
        }
      },
      "AuthUserSummary": {
        "type": "object",
        "required": ["id", "email", "name", "role"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "role": {
            "type": "string",
            "enum": ["owner", "vet", "admin"],
            "example": "owner"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["user", "token", "expiresIn"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserSummary"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJ1c2VyLWlkIiwiZXhwIjoxNzA5MDAwMDAwfQ.signature",
            "description": "JWT access token — include as Authorization: Bearer {token}"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Long-lived refresh token for obtaining new access tokens"
          },
          "expiresIn": {
            "type": "integer",
            "example": 604800,
            "description": "Access token lifetime in seconds (default: 7 days)"
          }
        }
      },
...
      "RefreshTokenResponse": {
        "type": "object",
        "required": ["token", "expiresIn"],
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expiresIn": {
            "type": "integer",
            "example": 604800
          }
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "Bad request — invalid input or missing required fields",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid input data",
                "details": {
                  "field": "email",
                  "issue": "Invalid email format"
                }
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required. Please log in."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "Server error. Please try again later."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}